+++ src/Dignite.Site.Public.HttpApi.Client/ClientProxies/Dignite/Site/Public/Seo/FeedDocumentPublicClientProxy.cs
@@ -1,0 +1,7 @@
+// This file is part of FeedDocumentPublicClientProxy, you can customize it here
+// ReSharper disable once CheckNamespace
+namespace Dignite.Site.Public.Seo;
+
+public partial class FeedDocumentPublicClientProxy
+{
+}